backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
email-validator==2.2.0
python-multipart==0.0.20
psycopg2-binary==2.9.10
pytest==8.4.1
httpx==0.28.1

